+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * Introduce hook mechanism for additional customization of workflows
 * Sentence-level negative log-likelihood criterion for sequence tagging
 * '-' stands for stdin for inference tools (translate, lm, tag)
-
+* Optional source features per request (for domain control) with rest translation server
 ### Fixes and improvements
 
 * Integrate hooks with rest translation server